requirements.txt
python-dotenv
requests
fastapi
uvicorn 

python-multipart
pillow
cloudinary
sentence-transformers
numpy==1.26.4
yt_dlp
imageio-ffmpeg
gdown
httpx
websocket-client
